Use Dist::Zilla to manage releases
I am unsure what the protocol for authorship credits in co-maintained modules is. Since this version will be released to CPAN under my ID, I have put myself as the author.

The repo urls point to the original GitHub repos, rather than my fork.

If this is not the correct way to do this, please let me know.